backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.5.2
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.2 